Add test for Verifier

+++ eth-contracts/contracts/Verifier.sol
@@ -577,18 +577,13 @@ contract Verifier {
         return 0;
     }
     event Verified(string s);
-    function verifyTx(
-            uint[2] memory a,
-            uint[2][2] memory b,
-            uint[2] memory c,
-            uint[2] memory input
-        ) public returns (bool r) {
+    function verifyTx(uint256[2] memory a, uint256[2][2] memory b, uint256[2] memory c, uint256[2] memory input) public returns (bool r) {
         Proof memory proof;
         proof.A = Pairing.G1Point(a[0], a[1]);
         proof.B = Pairing.G2Point([b[0][0], b[0][1]], [b[1][0], b[1][1]]);
         proof.C = Pairing.G1Point(c[0], c[1]);
-        uint[] memory inputValues = new uint[](input.length);
-        for(uint i = 0; i < input.length; i++){
+        uint256[] memory inputValues = new uint[](input.length);
+        for(uint256 i = 0; i < input.length; i++){
             inputValues[i] = input[i];
         }
         if (verify(inputValues, proof) == 0) {
